coralogix-dataprime-demo/requirements.txt
# Core dependencies
pydantic==2.8.2
python-dotenv==1.0.1

# OpenTelemetry
opentelemetry-api>=1.32.1
opentelemetry-sdk>=1.32.0
opentelemetry-semantic-conventions>=0.48b0
opentelemetry-exporter-otlp>=1.32.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-instrumentation-sqlite3>=0.48b0
opentelemetry-instrumentation-flask>=0.48b0
opentelemetry-instrumentation-requests>=0.48b0
opentelemetry-instrumentation-psycopg2>=0.48b0

# Utilities
structlog==24.2.0  # For structured logging
requests==2.31.0  # For HTTP requests

# Flask web interface
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Async HTTP
websockets>=12.0
aiohttp>=3.10.0

# Database and caching
redis==5.0.1  # Redis client for distributed caching
psycopg2-binary==2.9.9  # PostgreSQL adapter for product catalog